CodeCarbon/requirements.txt
codecarbon
pandas
numpy
scikit-learn
beautifulsoup4
pillow
requests